oct31_single_image_layout/ziyang8/requirements.txt
einops==0.8.0
huggingface_hub==0.26.1
imageio==2.34.1
matplotlib==3.9.2
numpy==2.1.2
open3d==0.18.0
Pillow==11.0.0
pillow_heif==0.20.0
pytorch3d==0.7.7
scipy==1.14.1
skimage==0.0
timm==1.0.11
tqdm==4.66.2
wandb==0.18.5
xformers==0.0.28.post3
